% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biol2022.R
\name{roster_biol2022}
\alias{roster_biol2022}
\title{Process Roster Data for BIOL2022}
\usage{
roster_biol2022(file_path)
}
\arguments{
\item{file_path}{Path to the Excel file containing the roster data.}
}
\value{
A dataframe in long format with roster assignments.
}
\description{
This function handles roster processing specific to BIOL2022 unit.
}
