% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biol1007.R
\name{roster_biol1007}
\alias{roster_biol1007}
\title{Process Roster Data for BIOL1007}
\usage{
roster_biol1007(file_path)
}
\arguments{
\item{file_path}{Path to the Excel file containing the roster data.}
}
\value{
A dataframe in long format with roster assignments.
}
\description{
This function handles roster processing specific to BIOL1007 unit.
}
